+++ packages/charts/src/index.ts
@@ -1,5 +1,5 @@
 export * from '@ant-design/plots';
-export * from '@ant-design/flowchart';
+export * from '@digiforce-cloud/flowdesigner';
 export * from '@ant-design/graphs';
 export * from '@ant-design/maps';