+++ packages/flowchart/src/components/toolbar/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { CanvasToolbar as XFlowToolbarPanel } from '@antv/xflow';
+import { CanvasToolbar as XFlowToolbarPanel } from '@digiforce-cloud/xflow';
 import { useToolbarConfig } from './util';
 import { FlowchartProps } from '../../interface';